src/requirements.txt
requests
numpy
matplotlib
opencv-python
scikit-learn
torch
torchvision
python-telegram-bot